Fix: Refacotred AppointmentController
Also moved AvailableSlots method into ScheduleController
