+++ HealthCareABApi/Controllers/AppointmentController.cs
@@ -37,33 +37,7 @@ namespace HealthCareABApi.Controllers
             {
                 return Unauthorized("User ID is missing a token");
             }
-
-            //Kollar om den valda tiden är tillgänglig
-            var caregiverAvailability = await _availabilityRepository.GetByCaregiverIdAsync(request.CaregiverId);
-            if (!caregiverAvailability.Any(a => a.AvailableSlots.Contains(request.AppointmentTime)))
-            {
-                return BadRequest("The selected slot is not available");
-            }
-
-            //Är den tillgänglig skapas ett nytt möte
-            var appointment = new Appointment
-            {
-                PatientId = userId,
-                CaregiverId = request.CaregiverId,
-                DateTime = request.AppointmentTime,
-                Status = AppointmentStatus.Scheduled
-            };
-
-            //sparar ner mötet till databasen
-            await _appointmentRepository.CreateAsync(appointment);
-
-            //Plockar bort tiden från vårdpersonalens kalender så det inte kan dubbelbokas.
-            var availability = caregiverAvailability.FirstOrDefault(a => a.AvailableSlots.Contains(request.AppointmentTime));
-            if (availability != null)
-            {
-                availability.AvailableSlots.Remove(request.AppointmentTime);
-                await _availabilityRepository.UpdateAsync(availability.Id, availability);
-            }
+            var appointment = await _appointmentService.BookAppointmentAsync(userId, request);
 
             //Om allt är ok returneras 200.
             return Ok(new
@@ -72,7 +46,6 @@ namespace HealthCareABApi.Controllers
                 appointmentId = appointment.Id,
                 appointmentTime = appointment.DateTime
             });
-
         }
 
         [Authorize]
@@ -94,26 +67,6 @@ namespace HealthCareABApi.Controllers
             }
 
             return Ok(appointmentDtos);
-        }
-
-        // Ny metod för att hämta alla tillgängliga tider
-        [Authorize]
-        [HttpGet("availableslots")]
-        public async Task<IActionResult> GetAvailableSlots()
-        {
-
-            var allAvailability = await _availabilityRepository.GetAllAsync();
-            if (allAvailability == null || !allAvailability.Any())
-            {
-                return NotFound("No available slots found");
-            }
-
-            var availableSlots = allAvailability.SelectMany(a => a.AvailableSlots).ToList();
-
-            return Ok(new AvailabilityDTO
-            {
-                AvailableSlots = availableSlots
-            });
         }
     }
 }
